refactor: 견스타그램 atoms 파일 변경으로 인한 내부 코드 경로 수정
